rearranged code, cleanup
