+++ main/adapters/httphelper/common.go
@@ -1,15 +1,20 @@
 package httphelper
 
 import (
+	"bytes"
+	"crypto/sha256"
+	"encoding/base64"
+	"encoding/hex"
+	"encoding/json"
 	"fmt"
-	"github.com/go-chi/chi"
-	"github.com/google/uuid"
-	log "github.com/sirupsen/logrus"
-	"github.com/ubirch/ubirch-client-go/main/vars"
 	"io/ioutil"
 	"net/http"
 	"strings"
 	"time"
+
+	"github.com/go-chi/chi"
+	"github.com/google/uuid"
+	log "github.com/sirupsen/logrus"
 )
 
 const (
@@ -19,7 +24,103 @@ const (
 	ReadTimeout           = 1 * time.Second  // maximum duration for reading the entire request -> low since we only expect requests with small content
 	WriteTimeout          = 99 * time.Second // time after which the connection will be closed if response was not written -> this should never happen
 	IdleTimeout           = 60 * time.Second // time to wait for the next request when keep-alives are enabled
+
+	UUIDKey          = "uuid"
+	OperationKey     = "operation"
+	VerifyPath       = "verify"
+	HashEndpoint     = "hash"
+	RegisterEndpoint = "register"
+
+	BinType  = "application/octet-stream"
+	TextType = "text/plain"
+	JSONType = "application/json"
+
+	HexEncoding = "hex"
+
+	HashLen = 32
 )
+
+type HTTPRequest struct {
+	ID   uuid.UUID
+	Auth string
+	Hash Sha256Sum
+}
+
+type Sha256Sum [HashLen]byte
+
+// GetHash returns the hash from the request body
+func GetHash(r *http.Request) (Sha256Sum, error) {
+	rBody, err := ReadBody(r)
+	if err != nil {
+		return Sha256Sum{}, err
+	}
+
+	if IsHashRequest(r) { // request contains hash
+		return getHashFromHashRequest(r.Header, rBody)
+	} else { // request contains original data
+		return getHashFromDataRequest(r.Header, rBody)
+	}
+}
+
+func getHashFromDataRequest(header http.Header, data []byte) (hash Sha256Sum, err error) {
+	switch ContentType(header) {
+	case JSONType:
+		data, err = GetSortedCompactJSON(data)
+		if err != nil {
+			return Sha256Sum{}, err
+		}
+		log.Debugf("sorted compact JSON: %s", string(data))
+
+		fallthrough
+	case BinType:
+		// hash original data
+		return sha256.Sum256(data), nil
+	default:
+		return Sha256Sum{}, fmt.Errorf("invalid content-type for original data: "+
+			"expected (\"%s\" | \"%s\")", BinType, JSONType)
+	}
+}
+
+func getHashFromHashRequest(header http.Header, data []byte) (hash Sha256Sum, err error) {
+	switch ContentType(header) {
+	case TextType:
+		if ContentEncoding(header) == HexEncoding {
+			data, err = hex.DecodeString(string(data))
+			if err != nil {
+				return Sha256Sum{}, fmt.Errorf("decoding hex encoded hash failed: %v (%s)", err, string(data))
+			}
+		} else {
+			data, err = base64.StdEncoding.DecodeString(string(data))
+			if err != nil {
+				return Sha256Sum{}, fmt.Errorf("decoding base64 encoded hash failed: %v (%s)", err, string(data))
+			}
+		}
+		fallthrough
+	case BinType:
+		if len(data) != HashLen {
+			return Sha256Sum{}, fmt.Errorf("invalid SHA256 hash size: "+
+				"expected %d bytes, got %d bytes", HashLen, len(data))
+		}
+
+		copy(hash[:], data)
+		return hash, nil
+	default:
+		return Sha256Sum{}, fmt.Errorf("invalid content-type for hash: "+
+			"expected (\"%s\" | \"%s\")", BinType, TextType)
+	}
+}
+
+// forwards response to sender
+func SendResponse(w http.ResponseWriter, resp HTTPResponse) {
+	for k, v := range resp.Header {
+		w.Header().Set(k, v[0])
+	}
+	w.WriteHeader(resp.StatusCode)
+	_, err := w.Write(resp.Content)
+	if err != nil {
+		log.Errorf("unable to write response: %s", err)
+	}
+}
 
 // wrapper for http.Error that additionally logs the error message to std.Output
 func Error(uid uuid.UUID, w http.ResponseWriter, err error, code int) {
@@ -44,7 +145,7 @@ func AuthToken(header http.Header) string {
 
 // getUUID returns the UUID parameter from the request URL
 func GetUUID(r *http.Request) (uuid.UUID, error) {
-	uuidParam := chi.URLParam(r, vars.UUIDKey)
+	uuidParam := chi.URLParam(r, UUIDKey)
 	id, err := uuid.Parse(uuidParam)
 	if err != nil {
 		return uuid.Nil, fmt.Errorf("invalid UUID: \"%s\": %v", uuidParam, err)
@@ -61,5 +162,36 @@ func ReadBody(r *http.Request) ([]byte, error) {
 }
 
 func IsHashRequest(r *http.Request) bool {
-	return strings.HasSuffix(r.URL.Path, vars.HashEndpoint)
+	return strings.HasSuffix(r.URL.Path, HashEndpoint)
+}
+
+func GetSortedCompactJSON(data []byte) ([]byte, error) {
+	var reqDump interface{}
+	var sortedCompactJson bytes.Buffer
+
+	// json.Unmarshal returns an error if data is not valid JSON
+	err := json.Unmarshal(data, &reqDump)
+	if err != nil {
+		return nil, fmt.Errorf("unable to parse JSON request body: %v", err)
+	}
+	// json.Marshal sorts the keys
+	sortedJson, err := jsonMarshal(reqDump)
+	if err != nil {
+		return nil, fmt.Errorf("unable to serialize json object: %v", err)
+	}
+	// remove spaces and newlines
+	err = json.Compact(&sortedCompactJson, sortedJson)
+	if err != nil {
+		return nil, fmt.Errorf("unable to compact json object: %v", err)
+	}
+
+	return sortedCompactJson.Bytes(), nil
+}
+
+func jsonMarshal(v interface{}) ([]byte, error) {
+	buffer := &bytes.Buffer{}
+	encoder := json.NewEncoder(buffer)
+	encoder.SetEscapeHTML(false)
+	err := encoder.Encode(v)
+	return buffer.Bytes(), err
 }
