+++ main/adapters/httphelper/http_server_test.go
@@ -1,4 +1,4 @@
-package handlers
+package httphelper
 
 import (
 	"bytes"
