+++ main/adapters/repository/init_db.go
@@ -8,7 +8,6 @@ import (
 	log "github.com/sirupsen/logrus"
 	"github.com/ubirch/ubirch-client-go/main/config"
 	"github.com/ubirch/ubirch-client-go/main/ent"
-	"github.com/ubirch/ubirch-client-go/main/vars"
 	"os"
 )
 
@@ -17,8 +16,8 @@ const MigrationVersion = "1.0.1"
 const (
 	PostgresIdentity = iota
 	PostgresVersion
-	//MySQL
-	//SQLite
+	PostgreSqlIdentityTableName string = "identity"
+	PostgreSqlVersionTableName  string = "version"
 )
 
 type Migration struct {
@@ -48,7 +47,7 @@ func Migrate(c config.Config) error {
 	txCtx, cancel := context.WithCancel(context.Background())
 	defer cancel()
 
-	dbManager, err := NewSqlDatabaseInfo(c.PostgresDSN, vars.PostgreSqlIdentityTableName)
+	dbManager, err := NewSqlDatabaseInfo(c.PostgresDSN, PostgreSqlIdentityTableName)
 	if err != nil {
 		return err
 	}
@@ -180,7 +179,7 @@ func checkVersion(ctx context.Context, dm *DatabaseManager) (*sql.Tx, bool, erro
 		return nil, false, err
 	}
 
-	if _, err := dm.db.Exec(CreateTable(PostgresVersion, vars.PostgreSqlVersionTableName)); err != nil {
+	if _, err := dm.db.Exec(CreateTable(PostgresVersion, PostgreSqlVersionTableName)); err != nil {
 		return tx, false, err
 	}
